cache each range request
