+++ aiocogeo/cog.py
@@ -5,7 +5,6 @@ from typing import Any, Callable, Dict, List, Optional, Tuple, Union
 from urllib.parse import urljoin
 import uuid
 
-from aiocache import cached, Cache
 import affine
 import numpy as np
 
@@ -19,14 +18,6 @@ from .partial_reads import PartialReadInterface
 logger = logging.getLogger(__name__)
 logger.setLevel(config.LOG_LEVEL)
 
-def config_cache(fn: Callable) -> Callable:
-    """
-    Inject cache config params (https://aiocache.readthedocs.io/en/latest/decorators.html#aiocache.cached)
-    """
-    def wrap_function(*args, **kwargs):
-        kwargs['cache_read'] = kwargs['cache_write'] = config.ENABLE_BLOCK_CACHE
-        return fn(*args, **kwargs)
-    return wrap_function
 
 @dataclass
 class COGReader(PartialReadInterface):
@@ -158,11 +149,7 @@ class COGReader(PartialReadInterface):
             )
         return gt
 
-    @config_cache
-    @cached(
-        cache=Cache.MEMORY,
-        key_builder=lambda fn,*args,**kwargs: f"{args[0].filepath}-{args[1]}-{args[2]}-{args[3]}"
-    )
+
     async def get_tile(self, x: int, y: int, z: int) -> np.ndarray:
 
         """
