+++ tests/test_cog_reader.py
@@ -289,10 +289,11 @@ async def test_block_cache_enabled(create_cog_reader, monkeypatch):
     infile = "https://async-cog-reader-test-data.s3.amazonaws.com/lzw_cog.tif"
     async with create_cog_reader(infile) as cog:
         await cog.get_tile(0,0,0)
-        request_count = cog.requests['count']
 
+    async with create_cog_reader(infile) as cog:
         await cog.get_tile(0,0,0)
-        assert cog.requests['count'] == request_count
+        # Confirm all requests are cached
+        assert cog.requests['count'] == 0
 
 
 @pytest.mark.asyncio
